fix: fix code

+++ src/me/bukkit/Infernaton/commands/DebugCommand.java
@@ -18,34 +18,30 @@ public class DebugCommand implements CommandExecutor {
 
     private FightToSurvive main;
 
-    public DebugCommand(FightToSurvive main){
+    public DebugCommand(FightToSurvive main) {
         this.main = main;
     }
 
     @Override
     public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
-        if (cmd.getName().equalsIgnoreCase("setPlayer")){
-            if (args.length == 0){
+        if (cmd.getName().equalsIgnoreCase("setPlayer")) {
+            if (args.length == 0) {
                 ChatHandler.sendCorrectUsage(sender, "Usage: /setPlayer <username>");
-            }
-            else if (args.length == 1){
+            } else if (args.length == 1) {
                 Player targetPlayer = Bukkit.getPlayerExact(args[0]);
                 System.out.println(targetPlayer);
-                if(targetPlayer == null){
+                if (targetPlayer == null) {
                     ChatHandler.sendError(sender, "Player not found");
-                }else{
+                } else {
                     main.HP().setPlayer(targetPlayer);
-                    ChatHandler.sendMessage(sender, "§r" +args[0] + " §fis ready to play !");
+                    ChatHandler.sendMessage(sender, "§r" + args[0] + " §fis ready to play !");
                 }
-            }
-            else {
+            } else {
                 ChatHandler.sendError(sender, "Too many argument");
             }
             return true;
-        }
-
-        else if (cmd.getName().equalsIgnoreCase("start")){
-            if (main.constH().isState(GState.WAITING)){
+        } else if (cmd.getName().equalsIgnoreCase("start")) {
+            if (main.constH().isState(GState.WAITING)) {
                 List<Player> redPlayers = main.constH().getRedTeam().getPlayers();
                 List<Player> bluePlayers = main.constH().getBlueTeam().getPlayers();
 
@@ -55,10 +51,10 @@ public class DebugCommand implements CommandExecutor {
                 //Compare if there the same numbers of players in each team
                 else if (redPlayers.size() == bluePlayers.size() && redPlayers.size() != 0) {
                     //Clear all players that attend to play
-                    for(Player player: main.constH().getBlueTeam().getPlayers()){
+                    for (Player player : main.constH().getBlueTeam().getPlayers()) {
                         main.HP().clear(player);
                     }
-                    for (Player player : main.constH().getRedTeam().getPlayers()){
+                    for (Player player : main.constH().getRedTeam().getPlayers()) {
                         main.HP().clear(player);
                     }
 
@@ -68,61 +64,56 @@ public class DebugCommand implements CommandExecutor {
                 } else {
                     ChatHandler.sendError(sender, "Not enough players.");
                 }
-            }else {
+            } else {
                 ChatHandler.sendError(sender, "Party already started !");
 
             }
             return true;
-        }
-
-        else if (cmd.getName().equalsIgnoreCase("cancelStart")){
-            if(main.constH().isState(GState.STARTING)){
+        } else if (cmd.getName().equalsIgnoreCase("cancelStart")) {
+            if (main.constH().isState(GState.STARTING)) {
                 main.constH().setState(GState.WAITING);
                 CountDown.stopAllCountdown(main);
-                ChatHandler.sendMessageListPlayer(main.constH().getAllTeamsPlayer(),"Launch canceled.");
-            }
-            else{
+                ChatHandler.sendMessageListPlayer(main.constH().getAllTeamsPlayer(), "Launch canceled.");
+            } else {
                 ChatHandler.sendError(sender, "Any countdown aren't set right now.");
             }
             return true;
-        }
-
-        else if (cmd.getName().equalsIgnoreCase("reset")){
-            if (main.constH().isState(GState.PLAYING)){
+        } else if (cmd.getName().equalsIgnoreCase("reset")) {
+            if (main.constH().isState(GState.PLAYING)) {
                 ChatHandler.sendMessageListPlayer(main.constH().getAllTeamsPlayer(), "Canceling the game");
                 List<Player> redPlayers = main.constH().getRedTeam().getPlayers();
                 List<Player> bluePlayers = main.constH().getBlueTeam().getPlayers();
 
-                for (Player player: redPlayers) {
+                for (Player player : redPlayers) {
                     main.HP().setPlayer(player);
                 }
-                for (Player player: bluePlayers) {
+                for (Player player : bluePlayers) {
                     main.HP().setPlayer(player);
                 }
                 main.constH().setState(GState.WAITING);
-            }
-            else {
+            } else {
                 ChatHandler.sendError(sender, "Any game is playing right now.");
             }
             return true;
-        }
-        else if (cmd.getName().equalsIgnoreCase("manage_time") && sender instanceof Player) {
+        } else if (cmd.getName().equalsIgnoreCase("manage_time") && sender instanceof Player) {
             final Player player = (Player) sender;
             DayNightCycle day = new DayNightCycle(main);
-        }
-        else if(cmd.getName().equalsIgnoreCase("getDoors")){
-           ChatHandler.sendInfoMessage(sender, "reset des portes...");
-            if ( main.constH().isState(GState.PLAYING)) {
+        } else if (cmd.getName().equalsIgnoreCase("getDoors")) {
+            ChatHandler.sendInfoMessage(sender, "reset des portes...");
+            if (main.constH().isState(GState.WAITING)) {
                 DoorListeners setDoors = new DoorListeners(this.main);
                 setDoors.setAllDoors();
+            } else {
+                ChatHandler.sendError(sender, "Can't perform this command while the game is pending");
             }
             return true;
-        }
-        else if(cmd.getName().equalsIgnoreCase("deleteDoors")){
+        } else if (cmd.getName().equalsIgnoreCase("deleteDoors")) {
             ChatHandler.sendInfoMessage(sender, "delete des portes..");
-            if ( main.constH().isState(GState.PLAYING)) {
-            DoorListeners setDoors = new DoorListeners(this.main);
-            setDoors.deleteAllDoors();
+            if (main.constH().isState(GState.WAITING)) {
+                DoorListeners setDoors = new DoorListeners(this.main);
+                setDoors.deleteAllDoors();
+            } else {
+                ChatHandler.sendError(sender, "Can't perform this command while the game is pending");
             }
             return true;
         }
